fix: Mobile UI Fix

+++ lib/models/current_weather.dart
@@ -155,7 +155,7 @@ class Main {
 }
 
 class Wind {
-  double? speed;
+  int? speed;
   int? deg;
   double? gust;
